Added qrcode field to event
